fix(runtime): use generated comparators for ignore-field-drift logging
The log-only drift detection in logIgnoredFieldDrift compared each ignored
field byte-for-byte. For fields the code-generator compares semantically --
is_iam_policy (IAMPolicyDocumentEqual) and is_document (DocumentEqual) -- AWS
canonicalizes the value on read (re-orders keys, collapses single-element
arrays, URL-encodes, reindents), so the stored declared string is never
byte-equal to the value read back. This produced a spurious "skipping drifted
ignore-field-drift fields" log line on every reconcile for any ignored
policy/document field, even when it never actually drifted.

Replace the byte comparison with driftedIgnoredPaths, which runs the generated
per-resource Delta against a copy of desired with the annotation removed (so
the delta's own FilterIgnoredDeltaDifferences is a no-op and ignored-path diffs
stay visible). This inherits each field's real comparator, so the log fires
only on genuine semantic drift. The stored CR is never mutated.

Verified end-to-end against the IAM Role AssumeRolePolicyDocument field: no
spurious log across 10 steady-state reconciles; the log now appears only on a
real out-of-band policy change.
